[Refactor] : weaksound Swagger 코드 추가 (#69)
