+++ src/main/java/com/potato/balbambalbam/config/SecurityConfig.java
@@ -2,10 +2,10 @@ package com.potato.balbambalbam.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.potato.balbambalbam.data.repository.RefreshRepository;
-import com.potato.balbambalbam.user.join.jwt.CustomLogoutFilter;
-import com.potato.balbambalbam.user.join.jwt.JWTFilter;
-import com.potato.balbambalbam.user.join.jwt.JWTUtil;
-import com.potato.balbambalbam.user.join.jwt.LoginFilter;
+import com.potato.balbambalbam.profile.join.jwt.CustomLogoutFilter;
+import com.potato.balbambalbam.profile.join.jwt.JWTFilter;
+import com.potato.balbambalbam.profile.join.jwt.JWTUtil;
+import com.potato.balbambalbam.profile.join.jwt.LoginFilter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
