+++ src/main/java/com/potato/balbambalbam/learningInfo/test/dto/WeakSoundTestDto.java
@@ -1,0 +1,13 @@
+package com.potato.balbambalbam.learningInfo.test.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Map;
+
+@Getter
+@Setter
+public class WeakSoundTestDto {
+    private Map<String, Integer> userWeakPhoneme; //초성, 중성
+    private Map<String, Integer> userWeakPhonemeLast; //종성
+}
