+++ src/main/java/com/potato/balbambalbam/main/cardList/controller/CardWeakSoundController.java
@@ -2,8 +2,8 @@ package com.potato.balbambalbam.main.cardList.controller;
 
 import com.potato.balbambalbam.exception.ExceptionDto;
 import com.potato.balbambalbam.main.cardList.service.CardWeakSoundService;
-import com.potato.balbambalbam.user.join.jwt.JWTUtil;
-import com.potato.balbambalbam.user.join.service.JoinService;
+import com.potato.balbambalbam.profile.join.jwt.JWTUtil;
+import com.potato.balbambalbam.profile.join.service.JoinService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
