+++ src/main/java/com/potato/balbambalbam/main/cardList/controller/CustomCardListController.java
@@ -4,8 +4,8 @@ import com.potato.balbambalbam.exception.ExceptionDto;
 import com.potato.balbambalbam.main.cardList.dto.CardListResponseDto;
 import com.potato.balbambalbam.main.cardList.dto.ResponseCardDto;
 import com.potato.balbambalbam.main.cardList.service.CardListService;
-import com.potato.balbambalbam.user.join.jwt.JWTUtil;
-import com.potato.balbambalbam.user.join.service.JoinService;
+import com.potato.balbambalbam.profile.join.jwt.JWTUtil;
+import com.potato.balbambalbam.profile.join.service.JoinService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
