+++ src/main/java/com/potato/balbambalbam/main/customCard/controller/CustomCardController.java
@@ -5,8 +5,8 @@ import com.potato.balbambalbam.exception.ExceptionDto;
 import com.potato.balbambalbam.main.customCard.dto.CustomCardRequestDto;
 import com.potato.balbambalbam.main.customCard.dto.CustomCardResponseDto;
 import com.potato.balbambalbam.main.customCard.service.CustomCardService;
-import com.potato.balbambalbam.user.join.jwt.JWTUtil;
-import com.potato.balbambalbam.user.join.service.JoinService;
+import com.potato.balbambalbam.profile.join.jwt.JWTUtil;
+import com.potato.balbambalbam.profile.join.service.JoinService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
