+++ src/main/java/com/potato/balbambalbam/profile/join/dto/CustomUserDetails.java
@@ -1,0 +1,56 @@
+package com.potato.balbambalbam.profile.join.dto;
+
+import com.potato.balbambalbam.data.entity.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Optional;
+
+public class CustomUserDetails  implements UserDetails {
+    private final User user;
+
+    public CustomUserDetails(Optional<User> optionalUser) {
+        this.user = optionalUser.orElse(null);
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Collection<GrantedAuthority> collection = new ArrayList<>();
+        if (user != null) {
+            collection.add(() -> user.getRole());
+        }
+        return collection;
+    }
+
+    @Override
+    public String getPassword() {
+        return "";
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getSocialId();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
