+++ src/main/java/com/potato/balbambalbam/profile/join/dto/DeleteUserDto.java
@@ -1,0 +1,12 @@
+package com.potato.balbambalbam.profile.join.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class DeleteUserDto {
+    @NotNull(message = "입력 데이터가 충분하지 않습니다.")
+    private String name;
+}
