+++ src/main/java/com/potato/balbambalbam/profile/join/jwt/CustomAuthenticationProvider.java
@@ -1,0 +1,40 @@
+package com.potato.balbambalbam.profile.join.jwt;
+
+import com.potato.balbambalbam.data.entity.User;
+import com.potato.balbambalbam.data.repository.UserRepository;
+import com.potato.balbambalbam.profile.join.dto.CustomUserDetails;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Component;
+
+import java.util.Optional;
+
+@Component
+public class CustomAuthenticationProvider implements AuthenticationProvider {
+    private final UserRepository userRepository;
+
+    public CustomAuthenticationProvider(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
+        String socialId = authentication.getName();
+        Optional<User> user = userRepository.findBySocialId(socialId);
+
+        if (user.isEmpty()) {
+            throw new UsernameNotFoundException("socialId에 맞는 user가 없습니다.");
+        }
+
+        CustomUserDetails userDetails = new CustomUserDetails(user);
+        return new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+    }
+
+    @Override
+    public boolean supports(Class<?> authentication) {
+        return UsernamePasswordAuthenticationToken.class.isAssignableFrom(authentication);
+    }
+}
