+++ src/main/java/com/potato/balbambalbam/review/controller/ReviewController.java
@@ -4,8 +4,8 @@ import com.potato.balbambalbam.exception.ExceptionDto;
 import com.potato.balbambalbam.review.dto.CardDto;
 import com.potato.balbambalbam.review.dto.ReviewListResponseDto;
 import com.potato.balbambalbam.review.service.ReviewService;
-import com.potato.balbambalbam.user.join.jwt.JWTUtil;
-import com.potato.balbambalbam.user.join.service.JoinService;
+import com.potato.balbambalbam.profile.join.jwt.JWTUtil;
+import com.potato.balbambalbam.profile.join.service.JoinService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
