--- src/main/java/com/potato/balbambalbam/user/join/dto/DeleteUserDto.java
@@ -1,12 +1,0 @@
-package com.potato.balbambalbam.user.join.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class DeleteUserDto {
-    @NotNull(message = "입력 데이터가 충분하지 않습니다.")
-    private String name;
-}
